Move escapeIdentifier and objectEntries into utils
