+++ src/utils.ts
@@ -1,4 +1,34 @@
 /**
+ * Escape the given identifier.
+ */
+export function escapeIdentifier(identifier: string) {
+  if (identifier.charAt(0) === '"' && identifier.charAt(identifier.length - 1) === '"') {
+    identifier = identifier.substr(1, identifier.length - 2)
+  }
+  if (identifier.includes('"')) {
+    throw new Error(`Invalid identifier: ${identifier}`)
+  }
+  return `"${identifier}"`
+}
+
+export type ValueRecord<T = any> = Record<string, T>
+
+/**
+ * Get the entries of the given object, optionally filtering out undefined values.
+ */
+export function objectEntries<Value = any>(
+  object: ValueRecord<Value>,
+  filterUndefinedValues?: boolean
+): Array<[string, Value]> {
+  const keys = Object.keys(object)
+  const entries = keys.map(key => [key, object[key]] as [string, Value])
+
+  return filterUndefinedValues
+    ? entries.filter(([, columnValue]) => typeof columnValue !== "undefined")
+    : entries
+}
+
+/**
  * Merge two lists, alternating between elements.
  *
  * @example
